move `setupTmpDirectory()` to functions.ts
